requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
langgraph>=0.0.10
langsmith>=0.0.69
python-dotenv==1.0.0
chromadb==0.4.22
beautifulsoup4==4.12.2
requests==2.31.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3 